* More 64-bit cleanup

+++ src/pl-funcs.h
@@ -822,7 +822,7 @@ COMMON(void)		wicPutNum(int64_t n, IOSTREAM *fd);
 COMMON(int64_t)		wicGetNum(IOSTREAM *fd);
 COMMON(void)		wicPutStringW(const pl_wchar_t *w, size_t len, 
 				      IOSTREAM *fd);
-COMMON(pl_wchar_t*)	wicGetStringUTF8(IOSTREAM *fd, unsigned *length,
+COMMON(pl_wchar_t*)	wicGetStringUTF8(IOSTREAM *fd, size_t *length,
 					 pl_wchar_t *buf, size_t bufsize);
 
 /* pl-write.c */
